Move SDK source files into sdk/ folder
