+++ sdk/EasyLink.cpp
