+++ sdk/easy_link_c.cpp
