+++ sdk/easy_link_c.h
