+++ src/CMakeLists.txt
@@ -1,8 +1,7 @@
-# Official SDK by Chessnut
-set(SDK_FILES EasyLink.h EasyLink.cpp easy_link_c.cpp easy_link_c.h)
-add_library(libeasylink SHARED ${SDK_FILES})
-add_library(libeasylink_static STATIC ${SDK_FILES})
-
 # Our own application
 add_executable (main main.c)
+
+# To find `easy_link_c.h` etc.; alternatively, "../sdk"
+target_include_directories(main PRIVATE "${CMAKE_SOURCE_DIR}/sdk")
+
 target_link_libraries (main libeasylink)
